% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_metric.R
\name{calc_dependency}
\alias{calc_dependency}
\title{Compare Dependency Metric Between Simulation Data And Observed Data}
\usage{
calc_dependency(sim_data, obs_data, pairs_matrix = NULL, percentile, sim_nr)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset. Rows correspond
to observations and columns correspond to covariate variables. sim_data
is supposed to include a column named "simulation_nr" as an identifier
of each simulation run.}

\item{obs_data}{A data.frame containing observation dataset. Rows correspond
to observations and columns correspond to covariates.}

\item{pairs_matrix}{Matrix with 2 column and each row containing a pair of
covariate names. If set to NULL, every possible covariate
pair is included.}

\item{percentile}{A numeric value representing the
percentile of the density contour for each pair combination of variables;
e.g., 90 represents that the overlap percentage is calculated for 90th percentile contours.}

\item{sim_nr}{An integer indicating the number of simulations.}
}
\value{
A data.frame containing the dependency metrics calculated for simulation data and observed data.
}
\description{
Compare Dependency Metric Between Simulation Data And Observed Data
}
\examples{
mtr_dependency <- calc_dependency(
    sim_data = pediatric_sim,
    obs_data = pediatric_3cov,
    pairs_matrix = NULL,
    percentile = 95,
    sim_nr = 100
    )

}
