% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_metric.R
\name{calc_margin}
\alias{calc_margin}
\title{Compare Marginal Metrics Between Simulation Data and Observed Data}
\usage{
calc_margin(
  sim_data,
  obs_data,
  sim_nr = NULL,
  var = NULL,
  aim_statistic = NULL
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset. Rows correspond
to observations and columns correspond to covariate variables. sim_data
is supposed to include a column named "simulation_nr" as an identifier
of each simulation run.}

\item{obs_data}{A data.frame containing observation dataset. Rows correspond
to observations and columns correspond to covariates.}

\item{sim_nr}{An integer indicating the number of simulations.}

\item{var}{A character vector containing the variables of interest for calculating
marginal metrics. If set to NULL, every covariate is included.}

\item{aim_statistic}{A character vector containing the marginal metrics of interest.
Available options include "mean", "median", "sd", "min", "max", "Q5.5%", and "Q95.95%".
If set to NULL, all marginal metrics are included.}
}
\value{
A data.frame containing the marginal metrics calculated for simulation data and observed data,
along with the differences between them.
}
\description{
Compare Marginal Metrics Between Simulation Data and Observed Data
}
\examples{
mtr_margin <- calc_margin(
    sim_data = pediatric_sim,
    obs_data = pediatric_3cov,
    sim_nr = 100,
    var = NULL,
    aim_statistic = c("mean", "median", "sd")
    )

}
