% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_metric.R
\name{calc_overlap}
\alias{calc_overlap}
\title{Calculate Overlap Percentage Of Simulation Data And Observed Data}
\usage{
calc_overlap(
  sim_data,
  obs_data,
  pairs_matrix = NULL,
  percentile,
  sim_nr,
  summarize = FALSE
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset. Rows correspond to observations and columns correspond to variables.}

\item{obs_data}{A data.frame containing observation dataset. Rows correspond to observations and columns correspond to variables.}

\item{pairs_matrix}{Matrix with 2 column and each row containing a pair of
variable names. If set to NULL, every possible variable
pair is included.}

\item{percentile}{A numeric value representing the
percentile of the density contour for each pair combination of variables;
e.g., 90 represents that the overlap percentage is calculated for 90th percentile contours.}

\item{sim_nr}{An integer indicating the number of simulations. If not set to 1, sim_data
is supposed to include a column named "simulation_nr" as an identifier of each simulation run.}

\item{summarize}{A logical value to indicate whether to summarize the overlap
percentage across different simulations using mean and standard deviation.}
}
\value{
A data.frame containing the overlap percentage of variable pair(s).
}
\description{
Calculate Overlap Percentage Of Simulation Data And Observed Data
}
\examples{
overlap <- calc_overlap(
    sim_data = pediatric_sim,
    obs_data = pediatric_3cov,
    pairs_matrix = matrix(c("AGE" ,"BW" ,"BW" ,"CREA" ), nrow = 2, ncol = 2),
    percentile = 95,
    sim_nr = 100,
    summarize = FALSE)

}
