% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC_dount_functions.R
\name{create_geom_donutVPC}
\alias{create_geom_donutVPC}
\title{Create ggplot layers for contours.}
\usage{
create_geom_donutVPC(
  sim_contours,
  conf_band = 95,
  colors_bands = c("#99E0DC", "#E498B4"),
  return_polygons = FALSE,
  show_median = FALSE
)
}
\arguments{
\item{sim_contours}{Output data.frame from the simulate_contours function.}

\item{conf_band}{A numeric value indicating the empirical confidence level for the width of the bands; e.g., 95 indicates 95% confidence interval.}

\item{colors_bands}{A vector with two strings specifying the colors of the confidence bands.}

\item{return_polygons}{A logical value to indicate whether return to a (list of) sf polygon object or a list containing the ggplot layers for a donut VPC.}
}
\value{
A list containing the ggplot layers for a donut VPC or a (list of) sf polygon object.
}
\description{
Create the ggplot contour layers for a donut VPC
}
