% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC_dount_functions.R
\name{create_polygon}
\alias{create_polygon}
\title{Extract the polygons from contour calculations.}
\usage{
create_polygon(ci_data)
}
\arguments{
\item{ci_data}{Output from extract_contour_df function, containing the
coordinates of contours, percentiles, density levels, covariate pairs and
simulation identifiers.}
}
\value{
A sf polygon object with the polygon(s) of a contour.
}
\description{
Create a sf polygon object from (multiple) contour(s)
}
