% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_example.R
\docType{data}
\name{df_example}
\alias{df_example}
\title{Pediatric population data}
\format{
A numeric data.frame
}
\source{
Virtual population generated from copula model \url{https://doi.org/10.1002/cpt.3099}
}
\usage{
df_example
}
\description{
A data.frame containing 3 covariates(age, body weight and serum creatine) of pediatric population.
}
\keyword{datasets}
