% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC_dount_functions.R
\name{extract_contour_df}
\alias{extract_contour_df}
\title{Restructure calculated contours in data.frame.}
\usage{
extract_contour_df(contour_list, contour_levels, sim_nr = NULL, pair = NULL)
}
\arguments{
\item{contour_list}{Output of the contourLines function, a list with contour(s)
of different density levels of interest.}

\item{contour_levels}{A numeric vector with density levels at which the
percentiles are calculated.}

\item{sim_nr}{An integer identifying the simulation for which contours are calculated.}

\item{pair}{A vector with two strings indicating the names of the covariates.}
}
\value{
A data.frame containing the coordinates of contours, percentiles, density levels,
covariate pairs and simulation identifiers.
}
\description{
Output from contourLines is converted from a list to a data.frame for
further use in other functions.
}
