% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC_dount_functions.R
\name{extract_geom_donutVPC}
\alias{extract_geom_donutVPC}
\title{Create ggplot layers for simulation data.}
\usage{
extract_geom_donutVPC(
  sim_data,
  percentiles,
  B,
  pairs_matrix = NULL,
  conf_band = 95,
  colors_bands = c("#99E0DC", "#E498B4"),
  return_polygons = FALSE
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset. A column called "simulation_nr" is reuqired to
indicate the number of simulations.}

\item{percentiles}{(A vector of) Numeric value(s) representing the
percentiles of the density distribution that appear in the VPC donut plot;
e.g., c(10, 50, 90) represents 10th, 50th and 90th percentiles.}

\item{B}{An integer indicating the number of simulations.}

\item{pairs_matrix}{Matrix with 2 column and each row containing a pair of
covariate names from the copula. If set to NULL, every possible covariate
pair is included.}

\item{conf_band}{A numeric value indicating the empirical confidence level for the width of the bands; e.g., 95 indicates 95\% confidence interval.}

\item{colors_bands}{A vector with two strings specifying the colors of the confidence bands.}

\item{return_polygons}{A logical value to indicate whether return to a (list of) sf polygon object or a list containing the ggplot layers for a donut VPC.}
}
\description{
Create ggplot layers for simulation data.
}
