% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_function.R
\name{get_overlap}
\alias{get_overlap}
\title{Calculate the overlap percentage of specific density contours of simulation data and observed data.}
\usage{
get_overlap(
  sim_data,
  obs_data,
  variables = NULL,
  percentile,
  B,
  summarize = FALSE
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset. Rows correspond to observations and columns correspond to variables.}

\item{obs_data}{A data.frame containing observation dataset. Rows correspond to observations and columns correspond to variables.}

\item{variables}{A character vector containing multiple characters. The
vector represents the variables of interest for calculating overlap. If
set to NULL, every covariate is included.}

\item{percentile}{A numeric value representing the
percentile of the density contour for each pair combination of variables;
e.g., 90 represents that the overlap percentage is calculated for 90th percentile contours.}

\item{B}{An integer indicating the number of simulations. If not set to 1, sim_data
is supposed to include a column named "simulation_nr" as an identifier of each simulation run.}

\item{summarize}{A logical value to indicate whether to summarize the overlap
percentage across different simulations using mean and standard deviation.}
}
\value{
A data.frame containing the overlap percentage of variable pair(s).
}
\description{
Calculate the overlap percentage of specific density contours of simulation data and observed data.
}
