% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repscore.R
\name{get_repscore}
\alias{get_repscore}
\title{Calculate the representation probability for a new (set of) patient(s)}
\usage{
get_repscore(obj, dat, type = "prob")
}
\arguments{
\item{obj}{an object of class `repscore` including the MC sample and their
corresponding densities}

\item{dat}{a data.frame containing the new patient data}

\item{type}{type of output, either the density, logdensity or the probability}
}
\value{
a numeric vector, with the same length as rows in dat. Wil contain either the density, log(density) or representation probability, depending on specified type.
}
\description{
Calculate the representation probability for a new (set of) patient(s)
}
