% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC_dount_functions.R
\name{ggVPC_donut}
\alias{ggVPC_donut}
\title{Create ggplot(s) with geom_vpc.}
\usage{
ggVPC_donut(geom_vpc, obs_data, pairs_data = NULL)
}
\arguments{
\item{geom_vpc}{Output list from the simulate_contours function, containing
the ggplot layers for a donut VPC.}

\item{obs_data}{A data.frame containing observation dataset.}

\item{save_path}{A string with path to save the resulting simulations as an R
object; if set to NULL, the object will be saved in the current directory.}
}
\value{
ggplot objects of donut VPC.
}
\description{
Create ggplot(s) with geom_vpc.
}
