% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_data.R
\docType{data}
\name{mix_data}
\alias{mix_data}
\title{A mixed dataset containing both discrete and continuous variables}
\format{
A numeric data.frame with 1000 records and 4 variables.
'Sex', character, category: male, female
'Age', numeric, range: 19 - 79 (year)
'Weight', numeric, range: 39.8 - 95.7 (kg)
'Albumin', numeric, range: 2.9 - 5.3 (g/dL)
}
\source{
Virtual population generated from copula model \url{https://doi.org/10.1002/cpt.3099}
}
\usage{
mix_data
}
\description{
A data.frame containing 4 covariates (Sex, Age, Weight and Albumin) of NHANES population.
}
\keyword{datasets}
