% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mVPC_function.R
\name{mVPC}
\alias{mVPC}
\title{Output Multivariate Visual Predictive Check Plot (mVPC)}
\usage{
mVPC(
  sim_data,
  obs_data,
  var = NULL,
  sim_nr,
  title = NULL,
  colors_bands = c("#99E0DC", "#E498B4"),
  caption = NULL,
  percentiles = c(10, 50, 90),
  conf_band = 95,
  return_grob = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset. Rows correspond
to observations and columns correspond to variables. sim_data
is supposed to include a column named "simulation_nr" as an identifier
of each simulation run.}

\item{obs_data}{A data.frame containing observation dataset. Rows correspond
to observations and columns correspond to variables.}

\item{var}{A character vector containing multiple characters. The
vector represents the variables of interest for generating qqplots. If
set to NULL, every variable is included.
#' @param sim_nr An integer indicating the number of simulations. Each simulation
generates the same number of individuals as the observation dataset.}

\item{title}{Title of the figure.}

\item{colors_bands}{A vector with two strings specifying the colors of the confidence bands.}

\item{caption}{A logical value to indicate define the size of caption.}

\item{percentiles}{(A vector of) Numeric value(s) representing the
percentiles of the density distribution that appear in the VPC donut plot;
e.g., c(10, 50, 90) represents 10th, 50th and 90th percentiles.}

\item{conf_band}{A numeric value indicating the empirical confidence level
for the width of the bands; e.g., 95 indicates 95\% confidence interval.}
}
\value{
Multivariate visual predictive check plot which contains qqplots and donutVPCs.
}
\description{
The multivariate visual predictive check plot (mVPC) combines donutVPCs with qqplots.
}
\examples{
plot <- mVPC(
    sim_data = pediatric_sim,
    obs_data = pediatric_3cov,
    var = NULL,
    sim_nr = 100,
    percentiles = c(10, 50, 90),
    colors_bands = c("#99E0DC", "#E498B4"))

}
