% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHANES_12cov.R
\docType{data}
\name{NHANES_12cov}
\alias{NHANES_12cov}
\title{NHANES population data}
\format{
A numeric data.frame
}
\source{
Virtual population generated from NHANES copula model \url{https://cocosim.lacdr.leidenuniv.nl/}
}
\usage{
NHANES_12cov
}
\description{
A data.frame containing 12 covariates (sex, race-ethnicity, age, body weight, height, fat mass, serum creatinine, ALT, AST, ALP, albumin and bilirubin) of the American population
}
\keyword{datasets}
