% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pediatric_sim.R
\docType{data}
\name{pediatric_sim}
\alias{pediatric_sim}
\title{Simulated pediatric population data}
\format{
A numeric data.frame with 44500 records and 4 variables.
'AGE', numeric, range: 0 - 736 (day)
'CREA', numeric, range: 0 - 158.5 ($\\mathrm{\\mu}$mol/L)
'BW', numeric, range: 0 - 18795.8 (g)
'simulation_nr', numeric, range: 1 - 100
}
\source{
Virtual population generated from pediatric_cop copula model (file path: system.file("extdata", "pediatric_cop.Rdata", package = "pmxcopula"))
}
\usage{
pediatric_sim
}
\description{
A data.frame containing 100 simulations of pediatric_3cov.
}
\keyword{datasets}
