% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repscore.R
\name{plot.repscore}
\alias{plot.repscore}
\title{Plot histogram from repscore MC sample}
\usage{
\method{plot}{repscore}(obj, vline = NULL)
}
\arguments{
\item{obj}{an object of class `repscore` including the MC sample and their
corresponding densities}

\item{vline}{probability value}
}
\description{
Plots a histogram of the densities of the log density from the MC sample. Uses ggplot2.
}
