% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_1dist.R
\name{plot_1dist}
\alias{plot_1dist}
\title{Plot Univariate Density Curves From Simulation Data And Observed Data}
\usage{
plot_1dist(
  sim_data,
  obs_data,
  sim_nr = NULL,
  var = NULL,
  pick_color = c("#3ABAC1", "#969696")
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset. Rows correspond
to observations and columns correspond to variables. sim_data
is supposed to include a column named "simulation_nr" as an identifier
of each simulation run.}

\item{obs_data}{A data.frame containing observation dataset. Rows correspond
to observations and columns correspond to variables.}

\item{sim_nr}{An integer indicating the number of simulations.Each simulation
generates the same number of individuals as the observation dataset.}

\item{var}{A character vector containing multiple characters. The
vector represents the variables of interest for generating qqplots. If
set to NULL, every variable is included.}

\item{pick_color}{A vector with two strings specifying the colors to represent the
simulaion and observation datasets.}
}
\value{
A ggplot object of density curves for variables.
}
\description{
Plot Univariate Density Curves From Simulation Data And Observed Data
}
\examples{
plot <- plot_1dist(
    sim_data = pediatric_sim,
    obs_data = pediatric_3cov,
    sim_nr = 100,
    pick_color = c("#3ABAC1","#969696"),
    var = NULL
    )

}
