% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donutVPC_function.R
\name{plot_donut}
\alias{plot_donut}
\title{Output donutVPCs from simulation data and observed data.}
\usage{
plot_donut(
  sim_data,
  obs_data,
  percentiles = c(10, 50, 90),
  sim_nr,
  pairs_matrix = NULL,
  conf_band = 95,
  colors_bands = c("#99E0DC", "#E498B4")
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset. Rows correspond
to observations and columns correspond to covariate variables. sim_data
is supposed to include a column named "simulation_nr" as an identifier
of each simulation run.}

\item{obs_data}{A data.frame containing observation dataset. Rows correspond
to observations and columns correspond to variables.}

\item{percentiles}{(A vector of) Numeric value(s) representing the
percentiles of the density distribution that appear in the VPC donut plot;
e.g., c(10, 50, 90) represents 10th, 50th and 90th percentiles.}

\item{sim_nr}{An integer indicating the number of simulations.}

\item{pairs_matrix}{Matrix with 2 column and each row containing a pair of
covariate names from the copula. If set to NULL, every possible covariate
pair is included.}

\item{conf_band}{A numeric value indicating the empirical confidence level
for the width of the bands; e.g., 95 indicates 95% confidence interval.}

\item{colors_bands}{A vector with two strings specifying the colors of the confidence bands.}
}
\value{
A patchwork object of donutVPC for covariate pairs.
}
\description{
Output donutVPCs from simulation data and observed data.
}
