% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mvdist.R
\name{plot_mvdist}
\alias{plot_mvdist}
\title{Visualize And Compare Dependency Structures Of Simulation Data And Observed Data}
\usage{
plot_mvdist(
  sim_data,
  obs_data,
  variables = NULL,
  sim_nr = 1,
  title = NULL,
  plot_type = "points",
  pick_color = c("#3ABAC1", "#969696"),
  full_plot = TRUE,
  caption = NULL,
  return_grob = FALSE
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset.}

\item{obs_data}{A data.frame containing observation dataset.}

\item{variables}{A character vector containing multiple characters. The
vector represents the variables of interest for plotting.
If set to NULL, every variable is included.}

\item{sim_nr}{An integer indicating the simulation data to use. One simulation
generates the same number of individuals as the observation dataset. If there are
multiple simulations or a column named "simulation_nr" in the simulation
dataset, select one specific simulation to visualize the distribution; if such column does not exist,
the default value of 1 will be used to visualize the entire silumation dataset.}

\item{title}{Title of the figure.}

\item{plot_type}{A character vector of plotting types. Either "density",
"points", or "both".}

\item{pick_color}{A vector with two strings specifying the colors to represent the
simulaion and observation datasets.}

\item{full_plot}{A logical value to indicate whether to plot multiple figures
on one page.}

\item{caption}{A logical value to indicate define the size of caption.}
}
\value{
A graphical object.
}
\description{
Visualize And Compare Dependency Structures Of Simulation Data And Observed Data
}
\examples{
plot <- plot_mvdist(
    sim_data = pediatric_sim,
    obs_data = pediatric_3cov,
    sim_nr = 1,
    pick_color = c("#3ABAC1","#969696"),
    plot_type = "density"
    )

}
