% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot_function.R
\name{plot_qq}
\alias{plot_qq}
\title{Output qqplots from simulation data and observed data.}
\usage{
plot_qq(
  sim_data,
  obs_data,
  sim_nr = NULL,
  conf_band = 95,
  var = NULL,
  type = "ribbon"
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset. Rows correspond
to observations and columns correspond to covariate variables. sim_data
is supposed to include a column named "simulation_nr" as an identifier
of each simulation run.}

\item{obs_data}{A data.frame containing observation dataset. Rows correspond
to observations and columns correspond to variables.}

\item{sim_nr}{An integer indicating the number of simulations.}

\item{conf_band}{A numeric value indicating the empirical confidence level
for the width of the bands; e.g., 95 indicates 95% confidence interval.}

\item{var}{A character vector containing multiple characters. The
vector represents the variables of interest for generating qqplots. If
set to NULL, every covariate is included.}

\item{type}{A character of plotting types. Either "ribbon" or"point".
"ribbon" corresponds to visualizing the confidence level defined in "conf_band" as shaded area,
and "point" corresponds to visualizing all the simulation quantiles as individual points.}
}
\value{
A patchwork object of qqplots for covariates.
}
\description{
Output qqplots from simulation data and observed data.
}
