% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot_sim_obs_distribution}
\alias{plot_sim_obs_distribution}
\title{Visualize and compare the dependency structures of simulation data and observed data.}
\usage{
plot_sim_obs_distribution(
  sim_data,
  obs_data,
  variables = NULL,
  sim_nr = 1,
  title = NULL,
  plot_type = "points",
  pick_color = c("#3ABAC1", "#969696"),
  full_plot = TRUE,
  caption = NULL
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset.}

\item{obs_data}{A data.frame containing observation dataset.}

\item{variables}{A character vector containing multiple characters. The
vector represents the variables of interest for plotting.
If set to NULL, every covariate is included.}

\item{sim_nr}{An integer indicating the number of simulations.
In each simulation, the amount of simulated individuals equals the amount of
individuals in observation dataset.}

\item{title}{Title of the figure.}

\item{plot_type}{A character vector of plotting types. Either "density",
"points", or "both".}

\item{pick_color}{A vector with two strings specifying the colors to represent the
simulaion and observation datasets.}

\item{full_plot}{A logical value to indicate whether to plot multiple figures
on one page.}

\item{caption}{A logical value to indicate define the size of caption.}
}
\value{
A graphical object.
}
\description{
Visualize and compare the dependency structures of simulation data and observed data.
}
