% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repscore.R
\name{print.repscore}
\alias{print.repscore}
\title{Print method for repscore objects}
\usage{
\method{print}{repscore}(obj)
}
\arguments{
\item{obj}{an object of class `repscore` including the MC sample and their
corresponding densities}
}
\description{
Print method for repscore objects
}
