% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function.R
\name{rcopula}
\alias{rcopula}
\title{Multiple Simulations For Copula Model Evaluation}
\usage{
rcopula(copula = copula, sim_nr = NULL)
}
\arguments{
\item{copula}{An object of class "vine_dist" or "vine", from rvinecopulib
package.}

\item{sim_nr}{An integer indicating the number of simulations.}
}
\value{
A data.frame containing simulation dataset. Rows correspond
to observations and columns correspond to covariate variables. sim_data includes
a column named "simulation_nr" as an identifier.
}
\description{
Multiple Simulations For Copula Model Evaluation
}
\examples{
# load a copula model: pediatric_cop
file_path <- system.file("extdata", "pediatric_cop.Rdata", package = "pmxcopula")
load(file_path)
pediatric_sim <- rcopula(pediatric_cop, sim_nr = 100)
}
