% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repscore.R
\name{repscore_MC}
\alias{repscore_MC}
\title{Monte Carlo sampling from vine object}
\usage{
repscore_MC(vine_obj, n, cores)
}
\arguments{
\item{vine_obj}{an object of class `vine_dist` or `vinecop_dist`.}

\item{n}{the number of MC samples.}

\item{cores}{the number of cores to use for parallel computations in vine
functions.}
}
\value{
an object of class `repscore` including the MC sample and their
corresponding densities
}
\description{
Samples from a vine copula object and calculates the MC density. The result
is a `repscore` object including the sample and the densities
}
