% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donutVPC_function.R
\name{donutVPC}
\alias{donutVPC}
\title{Output donutVPCs From Simulation Data And Observed Data}
\usage{
donutVPC(
  sim_data,
  obs_data,
  percentiles = c(10, 50, 90),
  sim_nr,
  pairs_matrix = NULL,
  conf_band = 95,
  colors_bands = c("#99E0DC", "#E498B4"),
  cores = 1
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset. Rows correspond
to observations and columns correspond to variables. sim_data
is supposed to include a column named "simulation_nr" as an identifier
of each simulation run.}

\item{obs_data}{A data.frame containing observation dataset. Rows correspond
to observations and columns correspond to variables.}

\item{percentiles}{(A vector of) Numeric value(s) representing the
percentiles of the density distribution that appear in the VPC donut plot;
e.g., c(10, 50, 90) represents 10th, 50th and 90th percentiles.}

\item{sim_nr}{An integer indicating the number of simulations. Each simulation
generates the same number of individuals as the observation dataset.}

\item{pairs_matrix}{Matrix with 2 column and each row containing a pair of
variable names. If set to NULL, every possible variable
pair is included.}

\item{conf_band}{A numeric value indicating the empirical confidence level
for the width of the bands; e.g., 95 indicates 95\% confidence interval.}

\item{colors_bands}{A vector with two strings specifying the colors of the confidence bands.}

\item{cores}{An integer of cores to use; if more than 1, calculation of donutVPC is done in parallel.}
}
\value{
A patchwork object of donutVPC for variable pairs.
}
\description{
Output donutVPCs From Simulation Data And Observed Data
}
\examples{
# plot the donut VPC for the variable pairs CREA-AGE and CREA-BW
plot <- donutVPC(
    sim_data = pediatric_sim,
    obs_data = pediatric_3cov,
    percentiles = c(10, 50, 90),
    sim_nr = 100,
    pairs_matrix = matrix(c("CREA","CREA", "AGE", "BW"),2,2),
    conf_band = 95,
    colors_bands = c("#99E0DC", "#E498B4"),
    cores = 2)

}
