% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIMIC_5cov.R
\docType{data}
\name{MIMIC_5cov}
\alias{MIMIC_5cov}
\title{MIMIC population data}
\format{
A numeric data.frame
}
\source{
Virtual population generated from copula model \url{https://doi.org/10.1002/cpt.3099}
}
\usage{
MIMIC_5cov
}
\description{
A data.frame containing 5 covariates(age, body weight , albumin, bilirubin, and serum creatinine) of ICU patient population.
}
\keyword{datasets}
