% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pediatric_3cov.R
\docType{data}
\name{pediatric_3cov}
\alias{pediatric_3cov}
\title{Pediatric population data}
\format{
A numeric data.frame with 1000 records and 3 variables.
'AGE', numeric, range: 0 - 720 (day)
'CREA', numeric, range: 5.4 - 150.0 ($\\mathrm{\\mu}$mol/L)
'BW', numeric, range: 155 - 18664 (g)
}
\source{
Virtual population generated from copula model \url{https://doi.org/10.1002/cpt.3099}
}
\usage{
pediatric_3cov
}
\description{
A data.frame containing 3 covariates(age, body weight and serum creatinine) of pediatric population.
}
\keyword{datasets}
