% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pediatric_sim.R
\docType{data}
\name{pediatric_sim}
\alias{pediatric_sim}
\title{Simulated pediatric population data}
\format{
A numeric data.frame
}
\source{
Virtual population generated from pediatric_cop copula model
}
\usage{
pediatric_sim
}
\description{
A data.frame containing 100 simulations of pediatric_3cov.
}
\keyword{datasets}
