% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot_function.R
\name{plot_qq}
\alias{plot_qq}
\title{Output qqplots From Simulation Data And Observed Data}
\usage{
plot_qq(
  sim_data,
  obs_data,
  sim_nr = NULL,
  conf_band = 95,
  var = NULL,
  type = "ribbon",
  qq_color = "lightblue"
)
}
\arguments{
\item{sim_data}{A data.frame containing simulation dataset. Rows correspond
to observations and columns correspond to covariate variables. sim_data
is supposed to include a column named "simulation_nr" as an identifier
of each simulation run.}

\item{obs_data}{A data.frame containing observation dataset. Rows correspond
to observations and columns correspond to variables.}

\item{sim_nr}{An integer indicating the number of simulations.}

\item{conf_band}{A numeric value indicating the empirical confidence level
for the width of the bands; e.g., 95 indicates 95\% confidence interval.}

\item{var}{A character vector containing multiple characters. The
vector represents the variables of interest for generating qqplots. If
set to NULL, every covariate is included.}

\item{type}{A character of plotting types. Either "ribbon" or"point".
"ribbon" corresponds to visualizing the confidence level defined in "conf_band" as shaded area,
and "point" corresponds to visualizing all simulation quantiles as individual points.}

\item{qq_color}{A character of color.}
}
\value{
A patchwork object of qqplots for variables.
}
\description{
Output qqplots From Simulation Data And Observed Data
}
\examples{
plot <- plot_qq(
    sim_data = pediatric_sim,
    obs_data = pediatric_3cov,
    sim_nr = 100,
    conf_band = 95,
    var = NULL,
    type = "ribbon"
    )
}
