% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC_dount_functions.R
\name{simulate_contours}
\alias{simulate_contours}
\title{Calculate density contours from simulation data.}
\usage{
simulate_contours(sim_data, percentiles, sim_nr, pairs_matrix = NULL)
}
\arguments{
\item{sim_data}{A data.frame containing simulation datasets with a column named 'sim_nr' to indicate the times of simulation.}

\item{percentiles}{(A vector of) Numeric value(s) representing the
percentiles of the density distribution that appear in the VPC donut plot;
e.g., c(10, 50, 90) represents 10th, 50th and 90th percentiles.}

\item{sim_nr}{An integer identifying the simulation for which contours are calculated.}

\item{pairs_matrix}{Matrix with 2 column and each row containing a pair of
covariate names from the copula. If set to NULL, every possible covariate
pair is included.}
}
\value{
A data.frame containing contours over all percentiles from the
different simulations .
}
\description{
Compute the contours from simulation data
}
